refactor: rename command `speed` to `perf`
