+++ src/commands/perf/index.ts
